Rewrite zoomIn and zoomOut functions; still tied to draw loop
